refactor(core): Move generateEmbedding to BaseLlmClient (#8442)
